Move progress to jobs package
